% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsbm.r
\name{ASE}
\alias{ASE}
\title{Adjacency spectral embedding and Laplacian embedding}
\usage{
ASE(A, d, lapace = "FALSE")
}
\arguments{
\item{A}{the adjacency matrix (double)}

\item{d}{the dimension of the embedding space (numeric)}

\item{lapace}{whether to obtain laplacian matrix or not (logical)}
}
\value{
the d dimension spectral embedding of the data
}
\description{
The function can get the spectral of the adjacency matrix and laplacian matrix
}
\examples{
\dontrun{
C<-matrix(c(0,1,1,1,0,0,1,0,0),nrow = 3)
ASE(C,2,lapace = TRUE)
}
}
