% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitsbm.r
\name{generater_model}
\alias{generater_model}
\title{generate SBM model}
\usage{
generater_model(
  B = matrix(c(0.2, 0.3, 0.3, 0.1), nrow = 2),
  p = c(0.3, 0.7),
  n = 1000
)
}
\arguments{
\item{B}{the probability matrix (double)}

\item{p}{the probability of the node belonging to each clusters (numeric)}

\item{n}{the number of node (numeric)}
}
\value{
A the adjacency matrix of the model
}
\description{
generate a SBM model with some settings
}
\examples{
\dontrun{
C<-generater_model()
}
}
