\name{SA23204186-package}
\alias{SA23204186-package}
\docType{package}
\title{Fitting a stochastic block model}
\description{This package use the spectral embedding and Gaussian mixture model to fit the stochastic block model. It has three function ,” generater_model” can generate a SBM model with the setting input,” Adjacency spectral embedding and Laplacian embedding” can obtain a embedding space of the adjacency matrix, “fitsbm”  can fit sbm model using spectral clustering.}
\author{SA23204186
Maintainer:SA23204186<zzj01@mail.ustc.edu.cn>}
\keyword{package}
