% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_cpp_prob}
\alias{sample_cpp_prob}
\title{C++ implementation of R's sample function with probabilities}
\usage{
sample_cpp_prob(x, size, replace = FALSE, prob = numericVector())
}
\arguments{
\item{x}{NumericVector - The input vector from R}

\item{size}{int - The number of elements to sample}

\item{replace}{bool - Whether sampling is done with replacement}

\item{prob}{NumericVector - The probabilities of elements to sample}
}
\value{
NumericVector - The sampled elements
}
\description{
C++ implementation of R's sample function with probabilities
}
